getGeolocation() method extracted to GoogleService class
GoogleService and GoogleServiceTest classes added
