+++ src/test/java/com/weather/service/PublicServiceHelperTest.java
@@ -21,12 +21,10 @@ import org.mockito.invocation.InvocationOnMock;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.mockito.stubbing.Answer;
 import org.springframework.test.util.ReflectionTestUtils;
+import static org.mockito.AdditionalMatchers.and;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.weather.model.DarkSkyWrapper;
-import com.weather.model.Geolocation;
-import com.weather.model.Geolocation.Result;
-import com.weather.model.Geolocation.Result.Geometry;
 import com.weather.model.Geolocation.Result.Geometry.Location;
 import com.weather.model.OpenWeatherWrapper;
 import com.weather.model.DarkSkyWrapper.CurrentWeather;
@@ -39,6 +37,9 @@ import com.weather.repository.OpenWeatherApiRepository;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PublicServiceHelperTest {
+    
+    @Mock
+    GoogleService googleService;
     
     @Mock
     GoogleApiRepository googleRepo;
@@ -54,9 +55,13 @@ public class PublicServiceHelperTest {
     
     @Before
     public void setup() {
-        ReflectionTestUtils.setField(underTest, "OPEN_WEATHER_APPID", "3d33157b92c33f9c5afa4c9a37b13715");
+        ReflectionTestUtils.setField(
+                underTest, "OPEN_WEATHER_APPID", "3d33157b92c33f9c5afa4c9a37b13715");
         ReflectionTestUtils.setField(underTest, "OPEN_WEATHER_FORMAT", "json");
-        ReflectionTestUtils.setField(underTest, "GEOLOCATION_APPID", "AIzaSyC3uvCOYFAgPG3e33QFozetMUh69oRBxrg");
+        ReflectionTestUtils.setField(
+                underTest, "GEOLOCATION_APPID", "AIzaSyC3uvCOYFAgPG3e33QFozetMUh69oRBxrg");
+        ReflectionTestUtils.setField(
+                googleService, "GEOLOCATION_APPID", "AIzaSyC3uvCOYFAgPG3e33QFozetMUh69oRBxrg");
     }
     
     @Test
@@ -71,42 +76,6 @@ public class PublicServiceHelperTest {
             + "&APPID=3d33157b92c33f9c5afa4c9a37b13715&mode=json"));
     }
     
-    @Test
-    public void checkGetGeolocation() throws JsonProcessingException, IOException {
-        //GIVEN
-        String city = "Szeged";
-        Mockito.when(googleRepo.getInformationAndMapToObject(BDDMockito.contains(city)))
-            .then(new Answer<Geolocation>() {
-            
-            @Override
-            public Geolocation answer(InvocationOnMock invoke) {
-                Geolocation geo = new Geolocation();
-                Result result = new Result();
-                Geometry geometry = new Geometry();
-                Location location = new Location();
-                
-                ArrayList<Result> resultList = new ArrayList<>();
-                
-                location.setLatitude("46.2530102");
-                location.setLongitude("20.1414254");
-                geometry.setLocation(location);
-                result.setGeometry(geometry);
-                resultList.add(result);
-                geo.setResults(resultList);
-                
-                return geo;
-            }
-        });
-        //WHEN
-        Location locationRes = underTest.getGeolocation(city);
-        //THEN
-        assertThat(locationRes.getLatitude(), containsString("46.2530102"));
-        assertThat(locationRes.getLongitude(), containsString("20.1414254"));
-        
-        Mockito.verify(googleRepo, Mockito.times(1)).getInformationAndMapToObject(
-                BDDMockito.contains(city));
-    }
-
     @Test
     public void checkOpenWeatherProviderProcess() throws JsonProcessingException, IOException {
         //GIVEN
@@ -147,7 +116,7 @@ public class PublicServiceHelperTest {
         //GIVEN
         String city = "Budapest";
         
-        Mockito.when(darkSkyRepo.getInformationAndMapToObject(BDDMockito.contains(city)))
+        Mockito.when(darkSkyRepo.getInformationAndMapToObject(BDDMockito.anyString()))
             .then(new Answer<DarkSkyWrapper>() {
             
             @Override
@@ -160,11 +129,14 @@ public class PublicServiceHelperTest {
                 currentWeather.setTemperature("20");
                 hourlyWeather.setSummary("Sunny with some cloudy breaks.");
                 
+                wrapper.setCurrentWeather(currentWeather);
+                wrapper.setHourWeather(hourlyWeather);
+                
                 return wrapper;
             }
         });
         
-        Mockito.when(underTest.getGeolocation(city)).then(new Answer<Location>() {
+        Mockito.when(googleService.getGeolocation(city)).then(new Answer<Location>() {
 
             @Override
             public Location answer(InvocationOnMock invoke) {
@@ -175,37 +147,15 @@ public class PublicServiceHelperTest {
             }
         });
         
-        Mockito.when(googleRepo.getInformationAndMapToObject(BDDMockito.anyString()))
-            .then(new Answer<Geolocation>() {
-            
-            @Override
-            public Geolocation answer(InvocationOnMock invoke) {
-                Geolocation geo = new Geolocation();
-                Result result = new Result();
-                Geometry geometry = new Geometry();
-                Location location = new Location();
-                
-                ArrayList<Result> resultList = new ArrayList<>();
-                
-                location.setLatitude("46.2530102");
-                location.setLongitude("20.1414254");
-                geometry.setLocation(location);
-                result.setGeometry(geometry);
-                resultList.add(result);
-                geo.setResults(resultList);
-                
-                return geo;
-            }
-        });
         //WHEN
         Map<String, String> map = underTest.darkSkyProviderProcess(city);
         //THEN
-        assertThat(map.get("desc") , containsString("Windy with rain"));
-        assertThat(map.get("temp") , containsString("14"));
+        assertThat(map.get("desc") , containsString("Sunny with some cloudy breaks."));
+        assertThat(map.get("temp") , containsString("20"));
         assertThat(map.get("imgUrl") , is(""));
         
-        Mockito.verify(darkSkyRepo, Mockito.times(1)).getInformationAndMapToObject(
-                BDDMockito.contains("Budapest"));
+        Mockito.verify(darkSkyRepo, Mockito.times(1)).getInformationAndMapToObject
+            (and(BDDMockito.contains("47.4979"), BDDMockito.contains("47.4979")));
     }
     
 }
